pyproject.toml
[project]
name = "gigpanel"
version = "1.0.0"
dependencies = [
	"qasync",
	"aiohttp",
	"midibox",
	"python-poppler-qt5",
]

[project.optional-dependencies]
all = [
	"platformdirs",
]
box = [
	"numpy",
]

[project.scripts]
gigpanel = "gigpanel.__main__:main"

[tool.setuptools]
packages = ["gigpanel"]

[[tool.mypy.overrides]]
module = "popplerqt5,qasync,midibox.*"
ignore_missing_imports = true
